+++ src/primary/client/auth/login_proof.rs
@@ -3,7 +3,7 @@ use sha1::{Sha1};
 use async_trait::async_trait;
 use serde::{Serialize, Deserialize};
 
-use crate::{with_opcode};
+use crate::primary::macros::with_opcode;
 use crate::primary::client::Opcode;
 use crate::primary::crypto::srp::Srp;
 use crate::primary::types::{HandlerInput, HandlerOutput, HandlerResult};
