+++ src/primary/client/realm/ping.rs
@@ -1,9 +1,9 @@
 use async_trait::async_trait;
 
+use crate::primary::macros::with_opcode;
 use crate::primary::client::opcodes::Opcode;
 use crate::primary::types::{HandlerInput, HandlerOutput, HandlerResult};
 use crate::primary::traits::packet_handler::PacketHandler;
-use crate::with_opcode;
 
 with_opcode! {
     @world_opcode(Opcode::CMSG_PING)
