+++ src/primary/client/realm/ready_for_account_data_times.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::{with_opcode};
+use crate::primary::macros::with_opcode;
 use crate::primary::client::opcodes::Opcode;
 use crate::primary::types::{HandlerInput, HandlerOutput, HandlerResult};
 use crate::primary::traits::packet_handler::PacketHandler;
