+++ src/primary/client/realm/request_characters.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 
-use crate::{with_opcode};
+use crate::primary::macros::with_opcode;
 use crate::primary::client::opcodes::Opcode;
 use crate::primary::types::{
     HandlerInput,
