+++ src/primary/macros/mod.rs
@@ -1,1 +1,3 @@
 mod with_opcode;
+
+pub(crate) use with_opcode::with_opcode;
